Refactor shared output logic into OutputFormat trait and fix lz4_flex vulnerability
Extract duplicated file-iteration, name-suffixing, empty-columns check,
and seed-derivation logic from Parquet/CSV/JSON implementations into a
default `generate_from_config` method on the `OutputFormat` trait. Each
format now only implements `generate_file` with its format-specific
writing logic.

Also bump lz4_flex 0.11.5 → 0.11.6 via Cargo.lock to resolve the high
severity vulnerability (dependabot alert #18).
